requirements.txt
numpy
opencv-python
scikit-learn
imutils
dlib
pytorch
torchvision
torchaudio
cudatoolkit